fix: move pytest_addoption to root conftest so --exhaustive works from tests/
The --exhaustive and --sampled flags were registered in a subdirectory
conftest.py, which pytest doesn't load when collecting from the top-level
tests/ directory. CI runs from tests/ root, so the flags were unrecognized.
