+++ tests/backend/semantic_graph/conftest.py
@@ -8,22 +8,6 @@ from backend.model.semantic_graph import SemanticGraph
 from backend.semantic_graph.sympy_translator import latex_to_semantic_graph
 
 
-def pytest_addoption(parser):
-    parser.addoption(
-        "--exhaustive",
-        action="store_true",
-        default=False,
-        help="Run the full exhaustive cross-product (~504 combos). Used by CI.",
-    )
-    parser.addoption(
-        "--sampled",
-        action="store",
-        default="200",
-        metavar="N",
-        help="Sample N cases from the full cross-product (default: 200).",
-    )
-
-
 @pytest.fixture
 def parse():
     """Return a parser callable for use in tests.
